@@ -1,4 +1,3 @@
-from typing import Tuple
 
 import torch
 
@@ -14,7 +13,7 @@ class CalQLImpl(CQLImpl):
         policy_obs: TorchObservation,
         value_obs: TorchObservation,
         returns_to_go: torch.Tensor,
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
+    ) -> tuple[torch.Tensor, torch.Tensor]:
         values, log_probs = super()._compute_policy_is_values(
             policy_obs=policy_obs,
             value_obs=value_obs,
