@@ -1,4 +1,3 @@
-from typing import Dict
 
 import torch
 
@@ -64,7 +63,7 @@ class TD3Impl(DDPGImpl):
 
     def inner_update(
         self, batch: TorchMiniBatch, grad_step: int
-    ) -> Dict[str, float]:
+    ) -> dict[str, float]:
         metrics = {}
 
         metrics.update(self.update_critic(batch))
