@@ -1,8 +1,7 @@
 import dataclasses
-from typing import Any, Dict, Sequence
+from typing import Any, Protocol, Sequence
 
 import numpy as np
-from typing_extensions import Protocol
 
 from ..constants import ActionSpace
 from ..types import (
@@ -282,7 +281,7 @@ class EpisodeBase(Protocol):
         """
         raise NotImplementedError
 
-    def serialize(self) -> Dict[str, Any]:
+    def serialize(self) -> dict[str, Any]:
         r"""Returns serized episode data.
 
         Returns:
@@ -291,7 +290,7 @@ class EpisodeBase(Protocol):
         raise NotImplementedError
 
     @classmethod
-    def deserialize(cls, serializedData: Dict[str, Any]) -> "EpisodeBase":
+    def deserialize(cls, serializedData: dict[str, Any]) -> "EpisodeBase":
         r"""Constructs episode from serialized data.
 
         This is an inverse operation of ``serialize`` method.
@@ -362,7 +361,7 @@ class Episode:
     def compute_return(self) -> float:
         return float(np.sum(self.rewards))
 
-    def serialize(self) -> Dict[str, Any]:
+    def serialize(self) -> dict[str, Any]:
         return {
             "observations": self.observations,
             "actions": self.actions,
@@ -371,7 +370,7 @@ class Episode:
         }
 
     @classmethod
-    def deserialize(cls, serializedData: Dict[str, Any]) -> "Episode":
+    def deserialize(cls, serializedData: dict[str, Any]) -> "Episode":
         return cls(
             observations=serializedData["observations"],
             actions=serializedData["actions"],
