@@ -4,7 +4,7 @@ import enum
 import os
 import random
 import re
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Optional
 from urllib import request
 
 import gym
@@ -52,9 +52,13 @@ __all__ = [
 DATA_DIRECTORY = "d3rlpy_data"
 DROPBOX_URL = "https://www.dropbox.com/s"
 CARTPOLE_URL = f"{DROPBOX_URL}/uep0lzlhxpi79pd/cartpole_v1.1.0.h5?dl=1"
-CARTPOLE_RANDOM_URL = f"{DROPBOX_URL}/4lgai7tgj84cbov/cartpole_random_v1.1.0.h5?dl=1"  # pylint: disable=line-too-long
+CARTPOLE_RANDOM_URL = (
+    f"{DROPBOX_URL}/4lgai7tgj84cbov/cartpole_random_v1.1.0.h5?dl=1"  # pylint: disable=line-too-long
+)
 PENDULUM_URL = f"{DROPBOX_URL}/ukkucouzys0jkfs/pendulum_v1.1.0.h5?dl=1"
-PENDULUM_RANDOM_URL = f"{DROPBOX_URL}/hhbq9i6ako24kzz/pendulum_random_v1.1.0.h5?dl=1"  # pylint: disable=line-too-long
+PENDULUM_RANDOM_URL = (
+    f"{DROPBOX_URL}/hhbq9i6ako24kzz/pendulum_random_v1.1.0.h5?dl=1"  # pylint: disable=line-too-long
+)
 
 
 def get_cartpole(
@@ -62,7 +66,7 @@ def get_cartpole(
     transition_picker: Optional[TransitionPickerProtocol] = None,
     trajectory_slicer: Optional[TrajectorySlicerProtocol] = None,
     render_mode: Optional[str] = None,
-) -> Tuple[ReplayBuffer, gym.Env[NDArray, int]]:
+) -> tuple[ReplayBuffer, gym.Env[NDArray, int]]:
     """Returns cartpole dataset and environment.
 
     The dataset is automatically downloaded to ``d3rlpy_data/cartpole.h5`` if
@@ -116,7 +120,7 @@ def get_pendulum(
     transition_picker: Optional[TransitionPickerProtocol] = None,
     trajectory_slicer: Optional[TrajectorySlicerProtocol] = None,
     render_mode: Optional[str] = None,
-) -> Tuple[ReplayBuffer, gym.Env[NDArray, NDArray]]:
+) -> tuple[ReplayBuffer, gym.Env[NDArray, NDArray]]:
     """Returns pendulum dataset and environment.
 
     The dataset is automatically downloaded to ``d3rlpy_data/pendulum.h5`` if
@@ -193,7 +197,7 @@ def get_atari(
     sticky_action: bool = True,
     pre_stack: bool = False,
     render_mode: Optional[str] = None,
-) -> Tuple[ReplayBuffer, gym.Env[NDArray, int]]:
+) -> tuple[ReplayBuffer, gym.Env[NDArray, int]]:
     """Returns atari dataset and envrironment.
 
     The dataset is provided through d4rl-atari. See more details including
@@ -221,7 +225,7 @@ def get_atari(
         tuple of :class:`d3rlpy.dataset.ReplayBuffer` and gym environment.
     """
     try:
-        import d4rl_atari  # type: ignore
+        import d4rl_atari  # type: ignore # noqa
 
         env = gym.make(
             env_name,
@@ -273,7 +277,7 @@ def get_atari_transitions(
     sticky_action: bool = True,
     pre_stack: bool = False,
     render_mode: Optional[str] = None,
-) -> Tuple[ReplayBuffer, gym.Env[NDArray, int]]:
+) -> tuple[ReplayBuffer, gym.Env[NDArray, int]]:
     """Returns atari dataset as a list of Transition objects and envrironment.
 
     The dataset is provided through d4rl-atari.
@@ -307,7 +311,7 @@ def get_atari_transitions(
         environment.
     """
     try:
-        import d4rl_atari
+        import d4rl_atari  # noqa
 
         # each epoch consists of 1M steps
         num_transitions_per_epoch = int(1000000 * fraction)
@@ -390,7 +394,7 @@ def get_d4rl(
     trajectory_slicer: Optional[TrajectorySlicerProtocol] = None,
     render_mode: Optional[str] = None,
     max_episode_steps: int = 1000,
-) -> Tuple[ReplayBuffer, gym.Env[NDArray, NDArray]]:
+) -> tuple[ReplayBuffer, gym.Env[NDArray, NDArray]]:
     """Returns d4rl dataset and envrironment.
 
     The dataset is provided through d4rl.
@@ -417,14 +421,14 @@ def get_d4rl(
         tuple of :class:`d3rlpy.dataset.ReplayBuffer` and gym environment.
     """
     try:
-        import d4rl
+        import d4rl  # noqa
         from d4rl.locomotion.wrappers import NormalizedBoxEnv
         from d4rl.utils.wrappers import (
             NormalizedBoxEnv as NormalizedBoxEnvFromUtils,
         )
 
         env = gym.make(env_name)
-        raw_dataset: Dict[str, NDArray] = env.get_dataset()  # type: ignore
+        raw_dataset: dict[str, NDArray] = env.get_dataset()  # type: ignore
 
         observations = raw_dataset["observations"]
         actions = raw_dataset["actions"]
@@ -470,7 +474,7 @@ def get_minari(
     trajectory_slicer: Optional[TrajectorySlicerProtocol] = None,
     render_mode: Optional[str] = None,
     tuple_observation: bool = False,
-) -> Tuple[ReplayBuffer, gymnasium.Env[Any, Any]]:
+) -> tuple[ReplayBuffer, gymnasium.Env[Any, Any]]:
     """Returns minari dataset and envrironment.
 
     The dataset is provided through minari.
@@ -654,7 +658,7 @@ def get_dataset(
     transition_picker: Optional[TransitionPickerProtocol] = None,
     trajectory_slicer: Optional[TrajectorySlicerProtocol] = None,
     render_mode: Optional[str] = None,
-) -> Tuple[ReplayBuffer, gym.Env[Any, Any]]:
+) -> tuple[ReplayBuffer, gym.Env[Any, Any]]:
     """Returns dataset and envrironment by guessing from name.
 
     This function returns dataset by matching name with the following datasets.
